+++ OUDS/Core/Themes/Sosh/Sources/Values/SoshBrandColorRawTokens+Values.swift
@@ -11,23 +11,12 @@
 // Software description: A SwiftUI components library with code examples for Orange Unified Design System 
 //
 
-import Foundation
-import SwiftUI
 import OUDSTokensRaw
 
-// MARK: - Type aliases to keep grammar clear
+// ଘ( ･ω･)_/ﾟ･:*:･｡☆
+// [File to generate with the tokenator]
 
-/// In the global design system tool, the verbs of "primitive colors" are used, and at this level mean simply raw tokens of colors.
-public typealias ColorSoshPrimitiveToken = ColorRawToken
-
-// MARK: Primitive tokens
-
-/// This extension helps to add new raw tokens for this theme.
-/// Type aliases here are just for consistancy reasons.
-/// Such tokens are packed in a _Swift enum_ so as to gather them in one object with the suitable namespace and avoid to have just constants in nothing else
-///  (i.e. publicly accessible from everywhere). More optimized than _struct_.
-///  All colors of the Sosh brand are listed here.
-public enum SoshBrandColorRawTokens {
+extension SoshBrandColorRawTokens {
 
     // MARK: Primitive token - Colors - Sosh - Magenta